@@ -19,8 +19,7 @@
         "source":"https://github.com/contao-bootstrap/navbar"
     },
     "require":{
-        "php":">=5.6.0",
-        "contao-community-alliance/composer-plugin": "~3.0",
+        "php":">=7.1",
         "contao/core-bundle":"^4.4",
         "contao-bootstrap/core": "^2.0",
         "menatwork/contao-multicolumnwizard":"^3.3.4",
@@ -28,8 +27,8 @@
         "contao-community-alliance/meta-palettes":"^1.10"
     },
     "require-dev": {
-        "netzmacht/contao-build-tools": "~1.1",
-        "contao/manager-plugin": "^2.0"
+        "contao/manager-plugin": "^2.0",
+        "phpcq/all-tasks": "^1.2"
     },
     "conflict": {
         "netzmacht/contao-bootstrap": "*"
