@@ -1,10 +1,13 @@
 <?php
 
 /**
- * @package   contao-bootstrap
- * @author    David Molineus <david.molineus@netzmacht.de>
- * @license   LGPL 3+
- * @copyright 2013-2017 netzmacht creative David Molineus
+ * Contao Bootstrap Navbar.
+ *
+ * @package    contao-bootstrap
+ * @author     David Molineus <david.molineus@netzmacht.de>
+ * @copyright  2017 netzmacht David Molineus. All rights reserved.
+ * @license    LGPL 3.0
+ * @filesource
  */
 
 namespace ContaoBootstrap\Navbar;
