@@ -1,11 +1,13 @@
 <?php
 
 /**
+ * Contao Bootstrap Navbar.
+ *
  * @package    contao-bootstrap
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @copyright  2017 netzmacht David Molineus. All rights reserved.
+ * @license    LGPL 3.0
  * @filesource
- *
  */
 
 
@@ -82,7 +84,6 @@ class NavigationHelper
     public static function createForTemplate(FrontendTemplate $template)
     {
         return new static($template);
-
     }
 
     /**
